Rename: 프로젝트 관련 파일 위치 일부 수정(#3)
